requirements.txt
Django==4.2
djangorestframework==3.16.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.1
pillow==11.3.0
yookassa==3.6.0
environs==14.3.0
